Rendering fixes, copy config tool fix, and cleanup

+++ src/main/kotlin/mods/eln/sixnode/lampsocket/LampSocketProcess.kt
@@ -185,7 +185,7 @@ class LampSocketProcess(var element: LampSocketElement) : IProcess {
             element.preparePacketForClient(packet)
 
             try {
-                packet.writeByte(element.lightValue)
+                packet.writeInt(element.lightValue)
             } catch (e: IOException) {
                 e.printStackTrace()
             }
